feat: add utility functions for formatting tool response content and parameters in observability components
